Refactor transcode endpoint to a separate file
